Added "Notification" and "book borrowing" features.

+++ src/Course.java
@@ -1,15 +1,15 @@
 import java.util.ArrayList;
 
 public class Course {
+
         String courseName;
         String courseCode;
-        double credit;
-
+        float credit;
         Teacher teacher;
         ArrayList<String>topics;
         double[] quiz = new double[4];
-        double midMark;
-        double finalMark;
+        float midMark;
+        float finalMark;
         int totalClass;
         int attendance;
